requirements.txt
############################
# ✅ Core Web Framework
############################
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

############################
# ✅ Database & ORM
############################
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0          # ✅ Replaced psycopg2 for async support
psycopg2-binary==2.9.9    # ✅ Keep only if Alembic CLI migrations need sync driver

############################
# ✅ Authentication & Security
############################
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

############################
# ✅ AI & Machine Learning
############################
openai==1.3.7
tiktoken==0.5.2

############################
# ✅ Background Tasks (Celery + Redis)
############################
celery[redis]==5.3.4
redis==4.5.4

############################
# ✅ HTTP Client & API Calls
############################
httpx==0.25.2
requests==2.31.0

############################
# ✅ Data Validation
############################
pydantic==2.5.0
pydantic-settings==2.1.0

############################
# ✅ File & Media Processing
############################
Pillow==10.1.0
python-magic==0.4.27
moviepy==1.0.3

############################
# ✅ Payments
############################
stripe==7.8.0

############################
# ✅ Monitoring & Logging
############################
sentry-sdk[fastapi]==1.38.0

############################
# ✅ Timezones
############################
pytz==2023.3

############################
# ✅ Development Only (Remove for production)
############################
pytest==7.4.4
pytest-asyncio==0.23.2
